Update datatypes for Numpy 1.24

Numpy 1.24 deprecated

    np.bool
    np.int
    np.float

Attempting to run MedPy raised an AttributeError that these weren't installed (I found this for metrics.hd95) This replaces them with np.bool_, int, and float, respectively, so medpy can continue to work with the latest Numpy.

Casts to specific precision (e.g. np.float32) are unchanged

+++ medpy/filter/image.py
@@ -1,15 +1,15 @@
 # Copyright (C) 2013 Oskar Maier
-# 
+#
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
 # the Free Software Foundation, either version 3 of the License, or
 # (at your option) any later version.
-# 
+#
 # This program is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU General Public License for more details.
-# 
+#
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
@@ -106,7 +106,7 @@ def sls(minuend, subtrahend, metric = "ssd", noise = "global", signed = True,
     pn_cval : scalar, optional
         Value to fill past edges of input if `pn_mode` is 'constant'. Default
         is 0.0
-        
+
     Returns
     -------
     sls : ndarray
@@ -114,7 +114,7 @@ def sls(minuend, subtrahend, metric = "ssd", noise = "global", signed = True,
 
     References
     ----------
-    
+
     .. [1] Mattias P. Heinrich, Mark Jenkinson, Manav Bhushan, Tahreema Matin, Fergus V. Gleeson, Sir Michael Brady, Julia A. Schnabel
            MIND: Modality independent neighbourhood descriptor for multi-modal deformable registration
            Medical Image Analysis, Volume 16, Issue 7, October 2012, Pages 1423-1435, ISSN 1361-8415
@@ -122,35 +122,35 @@ def sls(minuend, subtrahend, metric = "ssd", noise = "global", signed = True,
     """
     minuend = numpy.asarray(minuend)
     subtrahend = numpy.asarray(subtrahend)
-    
+
     if numpy.iscomplexobj(minuend):
         raise TypeError('complex type not supported')
     if numpy.iscomplexobj(subtrahend):
-        raise TypeError('complex type not supported')  
-    
+        raise TypeError('complex type not supported')
+
     mshape = [ii for ii in minuend.shape if ii > 0]
     sshape = [ii for ii in subtrahend.shape if ii > 0]
     if not len(mshape) == len(sshape):
         raise RuntimeError("minuend and subtrahend must be of same shape")
     if not numpy.all([sm == ss for sm, ss in zip(mshape, sshape)]):
         raise RuntimeError("minuend and subtrahend must be of same shape")
-    
+
     sn_footprint = __make_footprint(minuend, sn_size, sn_footprint)
     sn_fshape = [ii for ii in sn_footprint.shape if ii > 0]
     if len(sn_fshape) != minuend.ndim:
         raise RuntimeError('search neighbourhood footprint array has incorrect shape.')
-    
+
     #!TODO: Is this required?
     if not sn_footprint.flags.contiguous:
         sn_footprint = sn_footprint.copy()
-    
-    # created a padded copy of the subtrahend, whereas the padding mode is always 'reflect'  
+
+    # created a padded copy of the subtrahend, whereas the padding mode is always 'reflect'
     subtrahend = pad(subtrahend, footprint=sn_footprint, mode=sn_mode, cval=sn_cval)
-    
+
     # compute slicers for position where the search neighbourhood sn_footprint is TRUE
     slicers = [[slice(x, (x + 1) - d if 0 != (x + 1) - d else None) for x in range(d)] for d in sn_fshape]
     slicers = [sl for sl, tv in zip(itertools.product(*slicers), sn_footprint.flat) if tv]
-    
+
     # compute difference images and sign images for search neighbourhood elements
     ssds = [ssd(minuend, subtrahend[slicer], normalized=True, signed=signed, size=pn_size, footprint=pn_footprint, mode=pn_mode, cval=pn_cval) for slicer in slicers]
     distance = [x[0] for x in ssds]
@@ -162,23 +162,23 @@ def sls(minuend, subtrahend, metric = "ssd", noise = "global", signed = True,
     if 'global' == noise:
         variance = variance.sum() / float(numpy.product(variance.shape))
     # variance[variance < variance_global / 10.] = variance_global / 10. #!TODO: Should I keep this i.e. regularizing the variance to be at least 10% of the global one?
-    
+
     # compute sls
     sls = [dist_sign * numpy.exp(-1 * (dist / variance)) for dist_sign, dist in zip(distance_sign, distance)]
-    
+
     # convert into sls image, swapping dimensions to have varying patches in the last dimension
     return numpy.rollaxis(numpy.asarray(sls), 0, minuend.ndim + 1)
 
 def ssd(minuend, subtrahend, normalized=True, signed=False, size=None, footprint=None, mode="reflect", cval=0.0, origin=0):
     r"""
     Computes the sum of squared difference (SSD) between patches of minuend and subtrahend.
-    
+
     Parameters
     ----------
     minuend : array_like
         Input array from which to subtract the subtrahend.
     subtrahend : array_like
-        Input array to subtract from the minuend.    
+        Input array to subtract from the minuend.
     normalized : bool, optional
         Whether the SSD of each patch should be divided through the filter size for
         normalization. Default is 'True'.
@@ -207,15 +207,15 @@ def ssd(minuend, subtrahend, normalized=True, signed=False, size=None, footprint
     cval : scalar, optional
         Value to fill past edges of input if `mode` is 'constant'. Default
         is 0.0
-        
+
     Returns
     -------
     ssd : ndarray
         The patchwise sum of squared differences between minuend and subtrahend.
     """
     convolution_filter = average_filter if normalized else sum_filter
-    output = numpy.float if normalized else minuend.dtype
-    
+    output = float if normalized else minuend.dtype
+
     if signed:
         difference = minuend - subtrahend
         difference_squared = numpy.square(difference)
@@ -224,7 +224,7 @@ def ssd(minuend, subtrahend, normalized=True, signed=False, size=None, footprint
     else:
         distance = convolution_filter(numpy.square(minuend - subtrahend), size=size, footprint=footprint, mode=mode, cval=cval, origin=origin, output=output)
         distance_sign = 1
-    
+
     return distance, distance_sign
 
 def average_filter(input, size=None, footprint=None, output=None, mode="reflect", cval=0.0, origin=0):
@@ -277,7 +277,7 @@ def average_filter(input, size=None, footprint=None, output=None, mode="reflect"
     """
     footprint = __make_footprint(input, size, footprint)
     filter_size = footprint.sum()
-    
+
     output = _get_output(output, input)
     sum_filter(input, footprint=footprint, output=output, mode=mode, cval=cval, origin=origin)
     output /= filter_size
@@ -333,26 +333,26 @@ def sum_filter(input, size=None, footprint=None, output=None, mode="reflect", cv
     scipy.ndimage.convolve : Convolve an image with a kernel.
     """
     footprint = __make_footprint(input, size, footprint)
-    slicer = [slice(None, None, -1)] * footprint.ndim 
+    slicer = [slice(None, None, -1)] * footprint.ndim
     return convolve(input, footprint[slicer], output, mode, cval, origin)
 
 def otsu (img, bins=64):
     r"""
     Otsu's method to find the optimal threshold separating an image into fore- and background.
-    
+
     This rather expensive method iterates over a number of thresholds to separate the
     images histogram into two parts with a minimal intra-class variance.
-    
+
     An increase in the number of bins increases the algorithms specificity at the cost of
     slowing it down.
-    
+
     Parameters
     ----------
     img : array_like
         The image for which to determine the threshold.
     bins : integer
         The number of histogram bins.
-        
+
     Returns
     -------
     otsu : float
@@ -360,54 +360,54 @@ def otsu (img, bins=64):
     """
     # cast bins parameter to int
     bins = int(bins)
-    
+
     # cast img parameter to scipy arrax
     img = numpy.asarray(img)
-    
+
     # check supplied parameters
     if bins <= 1:
         raise AttributeError('At least a number two bins have to be provided.')
-    
+
     # determine initial threshold and threshold step-length
     steplength = (img.max() - img.min()) / float(bins)
     initial_threshold = img.min() + steplength
-    
+
     # initialize best value variables
     best_bcv = 0
     best_threshold = initial_threshold
-    
+
     # iterate over the thresholds and find highest between class variance
     for threshold in numpy.arange(initial_threshold, img.max(), steplength):
         mask_fg = (img >= threshold)
         mask_bg = (img < threshold)
-        
+
         wfg = numpy.count_nonzero(mask_fg)
         wbg = numpy.count_nonzero(mask_bg)
-        
+
         if 0 == wfg or 0 == wbg: continue
-        
+
         mfg = img[mask_fg].mean()
         mbg = img[mask_bg].mean()
-        
+
         bcv = wfg * wbg * math.pow(mbg - mfg, 2)
-        
+
         if bcv > best_bcv:
             best_bcv = bcv
             best_threshold = threshold
-        
+
     return best_threshold
 
 def local_minima(img, min_distance = 4):
     r"""
     Returns all local minima from an image.
-    
+
     Parameters
     ----------
     img : array_like
         The image.
     min_distance : integer
         The minimal distance between the minimas in voxels. If it is less, only the lower minima is returned.
-    
+
     Returns
     -------
     indices : sequence
@@ -427,7 +427,7 @@ def local_minima(img, min_distance = 4):
 def resample(img, hdr, target_spacing, bspline_order=3, mode='constant'):
         """
         Re-sample an image to a new voxel-spacing.
-        
+
         Parameters
         ----------
         img : array_like
@@ -440,11 +440,11 @@ def resample(img, hdr, target_spacing, bspline_order=3, mode='constant'):
             The bspline order used for interpolation.
         mode : str
             Points outside the boundaries of the input are filled according to the given mode ('constant', 'nearest', 'reflect' or 'wrap'). Default is 'constant'.
-            
+
         Warning
         -------
         Voxel-spacing of input header will be modified in-place!
-            
+
         Returns
         -------
         img : ndarray
@@ -454,14 +454,14 @@ def resample(img, hdr, target_spacing, bspline_order=3, mode='constant'):
         """
         if isinstance(target_spacing, numbers.Number):
             target_spacing = [target_spacing] * img.ndim
-        
+
         # compute zoom values
         zoom_factors = [old / float(new) for new, old in zip(target_spacing, header.get_pixel_spacing(hdr))]
-    
+
         # zoom image
         img = zoom(img, zoom_factors, order=bspline_order, mode=mode)
-        
+
         # set new voxel spacing
         header.set_pixel_spacing(hdr, target_spacing)
-        
+
         return img, hdr
